requirements.txt
PySide6>=6.6.0
openai>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
